+++ src/Sulu/Bundle/AdminBundle/Resources/views/Admin/index.html.dist.twig
@@ -13,7 +13,7 @@
         <!-- Place favicon.ico and apple-touch-icon.png in the root directory -->
 
         {% block stylesheets %}
-        <link rel="stylesheet" href="/bundles/suluadmin/dist/app.min.1486117677550.css">
+        <link rel="stylesheet" href="/bundles/suluadmin/dist/app.min.1486389523353.css">
         {% endblock stylesheets %}
     </head>
     <body>
@@ -70,7 +70,7 @@
         </script>
 
         {% block javascripts %}
-        <script data-main="/bundles/suluadmin/dist/app.min.1486117677550" src="/bundles/suluadmin/js/vendor/husky/husky.min.js?v=1486117677550"></script>
+        <script data-main="/bundles/suluadmin/dist/app.min.1486389523353" src="/bundles/suluadmin/js/vendor/husky/husky.min.js?v=1486389523353"></script>
         {% endblock javascripts %}
     </body>
 </html>
