+++ src/Sulu/Bundle/AdminBundle/Resources/views/CsvExport/form.html.twig
@@ -1,4 +1,5 @@
 <div class="grid">
+    {% block content %}
     <div class="grid-row">
         <div class="grid-col-6">
             <div class="form-group">
@@ -75,4 +76,5 @@
             </div>
         </div>
     </div>
+    {% endblock %}
 </div>
