+++ src/Sulu/Bundle/AdminBundle/Resources/views/Security/login.html.dist.twig
@@ -14,7 +14,7 @@
     <!-- Place favicon.ico and apple-touch-icon.png in the root directory -->
 
     {% block stylesheets %}
-    <link rel="stylesheet" href="/bundles/suluadmin/dist/login.min.1486117677550.css">
+    <link rel="stylesheet" href="/bundles/suluadmin/dist/login.min.1486389523353.css">
 
     <link rel="stylesheet" type="text/css" href="/bundles/suluadmin/css/main.css"/>
     {% endblock stylesheets %}
@@ -48,7 +48,7 @@
     </script>
 
     {% block javascripts %}
-    <script data-main="/bundles/suluadmin/dist/login.min.1486117677550" src="/bundles/suluadmin/js/vendor/husky/husky.min.js?v=1486117677550"></script>
+    <script data-main="/bundles/suluadmin/dist/login.min.1486389523353" src="/bundles/suluadmin/js/vendor/husky/husky.min.js?v=1486389523353"></script>
     {% endblock javascripts %}
 </body>
 </html>
